Add `ScheduledThreadPool::with_name_and_drop_behavior` method that creates a `ScheduledThreadPool` with the provided name, number of threads, and the desired behavior in relation to pending scheduled executions when the pool is dropped.
This on drop behavior is specified using the new `OnPoolDropBehavior` enum.

The existing behavior (and the default) is specified as
`CompletePendingScheduled`. This means that all pending scheduled executions
will be run, but periodic actions will not be rescheduled once these have
completed.

The other option is `DiscardPendingScheduled`. As the name suggests, this
behavior means that any pending scheduled executions will be discarded (not
run).